MGT-1 Add vendor changes (#2898)
* Add vendor changes
- Add createInterfaces() to ensure interfaces are created prior to assigning to VRF
- Remove setNoPeerGroup / conditional statement
- Enabled BGP global AFISAFI to BGP configs (REF: https://github.com/openconfig/featureprofiles/pull/2420)
- Added 3s sleep for Port2 test to ensure AFTs update
- Add 2 deviations SetRoutePrefNative, TableConnectionsUnsupported
- Changed BatchReplace to BatchUpdate to ensure route redistribution policy remains.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* Remove Deviation for nativepreference
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* remove dev from metadata.textproto
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* Fixed typo in deviation function
“This code is a Contribution to the OpenConfig Feature Profiles project (“Work”) made under the Google Software Grant and Corporate Contributor License Agreement (“CLA”) and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia’s intellectual property are granted for any other purpose. This code is provided on an “as is” basis without any warranties of any kind.”